# project_back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Exercices",
        "version": "1.0"
    },
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtenir un utilisateur spécifique",
                "description": "Récupère les détails d'un utilisateur par son ID.",
                "operationId": "6620b347ef3d3d7828ab93a3e064c0e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Modifier un utilisateur",
                "description": "Met à jour les informations d'un utilisateur par son ID.",
                "operationId": "19534687f146e916116334926644f683",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "new_password"
                                    },
                                    "picture": {
                                        "type": "string",
                                        "example": "profile_image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": 200,
                                    "content": "Utilisateur modifié avec succès"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprimer un utilisateur",
                "description": "Supprime un utilisateur par son ID.",
                "operationId": "c02d4b20c36fa979a0112ed40123e73a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": 200,
                                    "content": "Utilisateur supprimé avec succès"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Enregistrer un utilisateur",
                "description": "Crée un nouvel utilisateur.",
                "operationId": "f1fde7f4a932e393f6042d1b92e65b0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}/picture": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Ajouter ou mettre à jour la photo de profil d'un utilisateur",
                "description": "Ajoute ou met à jour la photo de profil de l'utilisateur.",
                "operationId": "72fae9cfefe658f55f4c85fb430de605",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "picture": {
                                        "description": "Image de profil",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo de profil mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Connexion d'un utilisateur",
                "description": "Authentifie un utilisateur et génère un jeton d'authentification.",
                "operationId": "71d8adfdcf176b9e5ef0ade00011ce34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "user": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john.doe@example.com"
                                    },
                                    "token": "generated_token"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Exercice": {
                "title": "Exercice",
                "description": "Un modèle représentant un exercice d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de l'exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de l'exercice",
                        "type": "string",
                        "example": "Pompes"
                    },
                    "description": {
                        "description": "Description de l'exercice",
                        "type": "string",
                        "example": "Exercice de renforcement musculaire"
                    },
                    "series": {
                        "description": "Nombre de séries",
                        "type": "integer",
                        "example": 3
                    },
                    "repetitions": {
                        "description": "Nombre de répétitions",
                        "type": "integer",
                        "example": 12
                    },
                    "training_id": {
                        "description": "ID de la séance associée à cet exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date de création de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Training": {
                "title": "Training",
                "description": "Un modèle représentant une séance d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de la séance",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de la séance",
                        "type": "string",
                        "example": "Séance de musculation"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur propriétaire de la séance",
                        "type": "integer",
                        "example": 1
                    },
                    "exercices": {
                        "description": "Liste des exercices associés à la séance",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Exercice"
                        }
                    },
                    "created_at": {
                        "description": "Date de création de la séance",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de la séance",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Modèle représentant un utilisateur",
                "properties": {
                    "id": {
                        "description": "ID de l'utilisateur",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'utilisateur",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "description": "Rôle de l'utilisateur",
                        "type": "string",
                        "example": "admin"
                    },
                    "picture": {
                        "description": "URL de la photo de profil de l'utilisateur",
                        "type": "string",
                        "example": "https://example.com/profile.jpg"
                    },
                    "email_verified_at": {
                        "description": "Date de vérification de l'email",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date de création de l'utilisateur",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de l'utilisateur",
                        "type": "string",
                        "format": "date-time"
                    },
                    "trainings": {
                        "description": "Liste des séances d'entraînement associées à cet utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Training"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}